+++ src/app/components/forms/dynamic-form/service/terra-json-to-form-field.service.ts
@@ -9,6 +9,7 @@ import { TerraFormFieldSelectBox } from '../data/terra-form-field-select-box';
 import { TerraFormFieldCheckBox } from '../data/terra-form-field-check-box';
 import { TerraFormFieldBaseContainer } from '../data/terra-form-field-base-container';
 import { TerraFormFieldCategoryPicker } from '../data/terra-form-field-category-picker';
+import { TerraFormFieldMultiCheckBox } from '../data/terra-form-field-multi-check-box';
 
 /**
  * @author mfrank
@@ -57,6 +58,9 @@ export class TerraJsonToFormFieldService
                 break;
             case TerraControlTypeEnum.CATEGORY_PICKER:
                 formField = new TerraFormFieldCategoryPicker(formFieldKey, formFieldData.label, formFieldData.required, formFieldData.options);
+                break;
+            case TerraControlTypeEnum.MULTI_CHECK_BOX:
+                formField = new TerraFormFieldMultiCheckBox(formFieldKey, formFieldData.label, formFieldData.required, formFieldData.options);
                 break;
             case TerraControlTypeEnum.VERTICAL_CONTAINER:
                 formField = this.createContainerFormField(formFieldKey, formFieldData, TerraControlTypeEnum.VERTICAL_CONTAINER);