+++ src/app/terra-components.module.ts
@@ -126,6 +126,8 @@ import { TerraDynamicFormComponent } from './components/forms/dynamic-form/terra
 import { TerraDynamicSwitchComponent } from './components/forms/dynamic-form/dynamic-switch/terra-dynamic-switch.component';
 import { TerraDynamicFormService } from './components/forms/dynamic-form/service/terra-dynamic-form.service';
 import { TerraFormFieldControlService } from './components/forms/dynamic-form/service/terra-form-field-control.service';
+import { TerraMultiCheckBoxComponent } from './components/forms/multi-check-box/terra-multi-check-box.component';
+import { TerraMultiCheckBoxComponentExample } from './components/forms/multi-check-box/example/terra-multi-check-box.component.example';
 import { TerraCategoryPickerComponent } from './components/category-picker/terra-category-picker.component';
 
 @NgModule({
@@ -159,6 +161,7 @@ import { TerraCategoryPickerComponent } from './components/category-picker/terra
         TerraSplitViewComponent,
         TerraFilterComponent,
         TerraMultiSelectBoxComponent,
+        TerraMultiCheckBoxComponent,
         TerraDynamicComponentLoaderComponent,
         TerraDynamicModuleLoaderComponent,
         TerraTileBoxComponent,
@@ -201,6 +204,7 @@ import { TerraCategoryPickerComponent } from './components/category-picker/terra
         TerraTextInputComponentExample,
         TerraCheckboxComponentExample,
         TerraMultiSelectBoxComponentExample,
+        TerraMultiCheckBoxComponentExample,
         TerraRadioButtonComponentExample,
         TerraSelectBoxComponentExample,
         TerraSliderComponentExample,
@@ -253,6 +257,7 @@ import { TerraCategoryPickerComponent } from './components/category-picker/terra
         TerraSplitViewComponent,
         TerraFilterComponent,
         TerraMultiSelectBoxComponent,
+        TerraMultiCheckBoxComponent,
         TerraTileBoxComponent,
         TerraTileBoxPanelComponent,
         TerraSuggestionBoxComponent,
@@ -302,6 +307,7 @@ import { TerraCategoryPickerComponent } from './components/category-picker/terra
         TerraSplitViewComponent,
         TerraFilterComponent,
         TerraMultiSelectBoxComponent,
+        TerraMultiCheckBoxComponent,
         TerraDynamicComponentLoaderComponent,
         TerraDynamicModuleLoaderComponent,
         TerraTileBoxComponent,
@@ -344,6 +350,7 @@ import { TerraCategoryPickerComponent } from './components/category-picker/terra
         TerraTextInputComponentExample,
         TerraCheckboxComponentExample,
         TerraMultiSelectBoxComponentExample,
+        TerraMultiCheckBoxComponentExample,
         TerraRadioButtonComponentExample,
         TerraSelectBoxComponentExample,
         TerraSliderComponentExample,