+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 ### Feature
+* **terra-multi-check-box** new component equal to *terra-multi-select-box* but with correct ngModel two way data binding
+* **terra-multi-select-box** this component is now deprecated and will be removed in the next major release
 * **terra-category-picker** new component to comfortable select a category using the **terra-node-tree**
 
 <a name="2.1.58-zlk"></a>